Update rcS and gitignore. Restore prev master.
A lot of hacks were made to get the old version working with JorConsole.
These have been fixed in JorConsole and now we can return to an almost
unmodified version of the sysroot files.

The only changes are to restore the tty size settings and check if the
emulator has networking activated.